refactor(P2.4): move WITH-detection predicates into render_plan/plan_predicates.rs
Phase-2 §5.1 module move. Extracted the WITH-detection predicate group verbatim
from plan_builder_utils.rs to a new render_plan/plan_predicates.rs:
has_with_clause_in_tree / has_with_clause_in_graph_rel / plan_contains_with_clause
(old lines 2008-2098). `pub(crate) use` re-exports left in plan_builder_utils for
the two production-called names; has_with_clause_in_tree has no production caller
(only the P1.2 characterization tests) so the test module imports it directly
rather than via a re-export that the lib build would flag unused. Zero logic edits.

Scoped down from §5.1's premise: the band also listed the private fresh-scan /
with-exported alias walkers (find_fresh_table_scan_aliases_in_plan /
collect_fresh_scan_aliases / with_exported_aliases_in_branch), but those are
coupled to P2.5's cte_rewrite fns (update_graph_joins_cte_refs /
rewrite_logical_expr_cte_refs), so they ride with P2.5 rather than churn
visibility ahead of their callers (§8.3 no-drive-by). The plan_builder_helpers
second copy of has_with_clause_in_graph_rel (pub(super), different semantics) is
the flagged D-cluster duplicate — a separate dedup, left untouched.

§5.1 table + §9 checklist + PRIORITIES P-3 updated.

Gates: 1612 lib + 529 integration (incl. corpus_sweep + sql_golden) + 1 ratchet
+ 7 unit, 0 failures; byte-identical (moved-fn bodies diff-clean vs main, no
golden churn, ratchet net-zero); fmt/clippy clean.
